# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MK Digital Operations Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List repositories
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRepositoryList"
    post:
      operationId: createRepository
      tags: [repositories]
      summary: Create repository record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
    patch:
      operationId: updateRepository
      tags: [repositories]
      summary: Update repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
    delete:
      operationId: deleteRepository
      tags: [repositories]
      summary: Delete repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /repositories/{id}/restore:
    post:
      operationId: restoreRepository
      tags: [repositories]
      summary: Restore a soft-deleted repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "404":
          description: Not found

  # ─── Releases ─────────────────────────────────────────────────────────────
  /releases:
    get:
      operationId: listReleases
      tags: [releases]
      summary: List releases
      parameters:
        - name: applicationId
          in: query
          schema:
            type: integer
        - name: environment
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Releases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReleaseList"
    post:
      operationId: createRelease
      tags: [releases]
      summary: Create release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
  /releases/{id}:
    get:
      operationId: getRelease
      tags: [releases]
      summary: Get release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
    patch:
      operationId: updateRelease
      tags: [releases]
      summary: Update release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
    delete:
      operationId: deleteRelease
      tags: [releases]
      summary: Delete release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /releases/{id}/restore:
    post:
      operationId: restoreRelease
      tags: [releases]
      summary: Restore a soft-deleted release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "404":
          description: Not found

  # ─── Security ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Repository:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: ["string", "null"]
        defaultBranch:
          type: ["string", "null"]
        visibility:
          type: string
        language:
          type: ["string", "null"]
        openPullRequests:
          type: integer
        openIssues:
          type: integer
        lastCommitAt:
          type: ["string", "null"]
        applicationId:
          type: ["integer", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        secretsExposed:
          type: boolean
        lastScannedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required: [id, name, visibility, openPullRequests, openIssues, status, secretsExposed, createdAt, updatedAt]
    RepositoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
        defaultBranch:
          type: string
        visibility:
          type: string
        language:
          type: string
        openPullRequests:
          type: integer
        openIssues:
          type: integer
        lastCommitAt:
          type: string
        applicationId:
          type: integer
        status:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        secretsExposed:
          type: boolean
        lastScannedAt:
          type: string
      required: [name, visibility, status]
    RepositoryUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        defaultBranch:
          type: string
        visibility:
          type: string
        language:
          type: string
        openPullRequests:
          type: integer
        openIssues:
          type: integer
        lastCommitAt:
          type: string
        applicationId:
          type: integer
        status:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        secretsExposed:
          type: boolean
        lastScannedAt:
          type: string

    # ── Releases ──────────────────────────────────────────────────────────
    Release:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        applicationName:
          type: ["string", "null"]
        version:
          type: string
        environment:
          type: string
        status:
          type: string
        releaseDate:
          type: ["string", "null"]
        releasedBy:
          type: ["string", "null"]
        releaseNotes:
          type: ["string", "null"]
        rollbackAvailable:
          type: boolean
        approved:
          type: boolean
        approvedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required: [id, applicationId, version, environment, status, rollbackAvailable, approved, createdAt, updatedAt]
    ReleaseInput:
      type: object
      properties:
        applicationId:
          type: integer
        version:
          type: string
          minLength: 1
        environment:
          type: string
        status:
          type: string
        releaseDate:
          type: string
        releasedBy:
          type: string
        releaseNotes:
          type: string
        rollbackAvailable:
          type: boolean
        approved:
          type: boolean
        approvedBy:
          type: string
      required: [applicationId, version, environment, status]
    ReleaseUpdate:
      type: object
      properties:
        version:
          type: string
        environment:
          type: string
        status:
          type: string
        releaseDate:
          type: string
        releasedBy:
          type: string
        releaseNotes:
          type: string
        rollbackAvailable:
          type: boolean
        approved:
          type: boolean
        approvedBy:
          type: string

    # ── Security ──────────────────────────────────────────────────────────
...
    PaginatedRepositoryList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        total:
          type: integer
    PaginatedReleaseList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Release"
        total:
          type: integer
...